Add admin check, seperated bot and reactbot, fixed pbcooldown
