+++ scripts/catfact.py
@@ -25,4 +25,5 @@ def error_cleanup(error_code):
         response_obj.messages_to_send.append("An unknown error occured. Error code: " + error_code)
     return response_obj
 
-    
+def is_admin_command():
+    return False
