+++ scripts/coin.py
@@ -11,3 +11,6 @@ def coin(response):
     m = random.randrange(0, 2)
     response_obj.messages_to_send.append(l[m])
     return response_obj
+
+def is_admin_command():
+    return False
